+++ tbx/docmaker/+docmaker/md2xml.m
@@ -1,7 +1,7 @@
 function doc = md2xml( md )
 %md2xml  Convert Markdown to XML
 %
-%   x = docstar.md2xml(md) converts the Markdown md to the XML document x
+%   x = docmaker.md2xml(md) converts the Markdown md to the XML document x
 %   using the GitHub API at: https://docs.github.com/en/rest/markdown
 %
 %   Authenticated requests get a higher API rate limit.  To authenticate,